Add "-q" flag for quiet operation
It is very common to invoke assemblers or compilers from a scripted
build (e.g. Makefile), and in that case it is desirable to avoid
unnecessary console output so that warnings or errors are more apparent
to the user.  This patch adds a "-q" option to all three utilities for
this purpose.  Sample usage:

    OBJS := com/example/class_a.class \
            com/example/class_b.class \
            com/example/class_c.class

    $(OBJS): %.class: %.j
            @echo "  ASM    $@"
            @python assemble.py -q $<

    new.jar: $(OBJS)
            @echo "  JAR    $@"
            @jar cf $@ $^

This yields:

    $ make new.jar
      ASM    com/example/class_a.class
      ASM    com/example/class_b.class
      ASM    com/example/class_c.class
      JAR    new.jar
    $